Unpin Numpy and Scipy min versions, require Numpy >= 2.0.0
The string representation of most scalars changed in Numpy 2.0.0
affecting doctests. Update the doctests; then require Numpy>=2.0.0
to ensure that the version of Numpy that we use at runtime is no
older than the minimum version that we test with.

See https://github.com/astropy/astropy/issues/15095.

+++ ligo/skymap/distance.py
@@ -113,7 +113,7 @@ Test against numerical integral of pdf.
 For negative distances, it returns 0.
 
 >>> conditional_cdf(-1, distmu, distsigma, distnorm)
-0.0
+np.float64(0.0)
 
 For infinite postive distance, it returns 1 (provided that distnorm normalizes
 the distribution).
@@ -418,7 +418,7 @@ Examples
 For negative distances, it returns 0.
 
 >>> marginal_cdf(-1, prob, distmu, distsigma, distnorm)
-0.0
+np.float64(0.0)
 
 For infinite postive distance, it returns the sum of prob
 (provided that distnorm normalizes the distribution).
