+++ ligo/skymap/plot/angle.py
@@ -54,7 +54,7 @@ def wrapped_angle(a):
     --------
 
     >>> wrapped_angle(3 * np.pi)
-    3.141592653589793
+    np.float64(3.141592653589793)
     """
     return np.mod(a, 2 * np.pi)
 
@@ -66,6 +66,6 @@ def wrapped_angle_deg(a):
     --------
 
     >>> wrapped_angle_deg(540.)
-    180.0
+    np.float64(180.0)
     """
     return np.mod(a, 360)
